+++ integration-test/tests/graph_editor.rs
@@ -5,6 +5,10 @@
 use enso_integration_test::prelude::*;
 
 use approx::assert_abs_diff_eq;
+use enso_frp::future::FutureEvent;
+use enso_frp::io::mouse::PrimaryButton;
+use enso_gui::view::graph_editor;
+use enso_gui::view::graph_editor::component::node as node_view;
 use enso_gui::view::graph_editor::component::node::test_utils::NodeModelExt;
 use enso_gui::view::graph_editor::component::node::Expression;
 use enso_gui::view::graph_editor::GraphEditor;
@@ -13,6 +17,8 @@ use enso_gui::view::graph_editor::NodeId;
 use enso_gui::view::graph_editor::NodeSource;
 use enso_web::sleep;
 use ensogl::display::navigation::navigator::ZoomEvent;
+use ensogl::display::scene::test_utils::MouseExt;
+use ensogl::display::Scene;
 use ordered_float::OrderedFloat;
 use std::time::Duration;
 
@@ -152,7 +158,10 @@ async fn adding_node_with_add_node_button() {
     let (node_id, node_source, _) = add_node_with_add_node_button(&graph_editor, "1");
     assert!(node_source.is_none());
     assert_eq!(graph_editor.nodes().all.len(), INITIAL_NODE_COUNT + 3);
-    let node_position = graph_editor.model.get_node_position(node_id).expect("Node was not added");
+    let node_position = graph_editor.model.get_node_position(node_id).expect(
+        "Node was not
+added",
+    );
     let center_of_screen = scene.screen_to_scene_coordinates(Vector3::zeros());
     assert_abs_diff_eq!(node_position.x, center_of_screen.x, epsilon = 10.0);
     assert_abs_diff_eq!(node_position.y, center_of_screen.y, epsilon = 10.0);
@@ -204,4 +213,104 @@ fn add_node_with_add_node_button(
     let add_node_button = &graph_editor.model.add_node_button;
     let method = |_: &GraphEditor| add_node_button.click();
     add_node(graph_editor, expression, method)
+}
+
+#[wasm_bindgen_test]
+async fn mouse_oriented_node_placement() {
+    struct Case {
+        scene:             Scene,
+        graph_editor:      GraphEditor,
+        source_node:       Node,
+        mouse_position:    Vector2,
+        expected_position: Vector2,
+    }
+
+    impl Case {
+        fn run(&self) {
+            self.check_tab_key();
+            self.check_edge_drop();
+        }
+
+        fn check_searcher_opening_place(
+            &self,
+            added_node: FutureEvent<(NodeId, Option<NodeSource>, bool)>,
+        ) {
+            let (new_node_id, _, _) = added_node.expect();
+            let new_node_pos =
+                self.graph_editor.model.get_node_position(new_node_id).map(|v| v.xy());
+            assert_eq!(new_node_pos, Some(self.expected_position));
+            self.graph_editor.stop_editing();
+            assert_eq!(self.graph_editor.model.nodes.all.len(), 2);
+        }
+
+        fn check_tab_key(&self) {
+            self.scene.mouse.frp.position.emit(self.mouse_position);
+            let added_node = self.graph_editor.node_added.next_event();
+            self.graph_editor.start_node_creation();
+            self.check_searcher_opening_place(added_node);
+        }
+
+        fn check_edge_drop(&self) {
+            let port = self.source_node.view.model.output_port_shape().unwrap();
+            port.events.mouse_down.emit(PrimaryButton);
+            port.events.mouse_up.emit(PrimaryButton);
+            self.scene.mouse.frp.position.emit(self.mouse_position);
+            assert!(
+                self.graph_editor.has_detached_edge.value(),
+                "No detached edge after clicking port"
+            );
+            let added_node = self.graph_editor.node_added.next_event();
+            self.scene.mouse.click_on_background();
+            self.check_searcher_opening_place(added_node);
+        }
+    }
+
+    let test = IntegrationTestOnNewProject::setup().await;
+    let scene = &test.ide.ensogl_app.display.default_scene;
+    let graph_editor = test.graph_editor();
+    let gap_x = graph_editor.default_x_gap_between_nodes.value();
+    let gap_y = graph_editor.default_y_gap_between_nodes.value();
+    let min_spacing = graph_editor.min_x_spacing_for_new_nodes.value();
+
+    let InitialNodes { above, below } = InitialNodes::obtain_from_graph_editor(&graph_editor);
+
+    let create_case =
+        |source_node: &Node, mouse_position: Vector2, expected_position: Vector2| Case {
+            scene: scene.clone_ref(),
+            graph_editor: graph_editor.clone_ref(),
+            source_node: source_node.clone_ref(),
+            mouse_position,
+            expected_position,
+        };
+
+    let far_away = below.position().xy() + Vector2(500.0, 500.0);
+    let far_away_expect = far_away;
+    create_case(&below, far_away, far_away_expect).run();
+
+    let under_below = below.position().xy() + Vector2(30.0, -15.0);
+    let under_below_expect = below.position().xy() + Vector2(0.0, -gap_y - node_view::HEIGHT);
+    create_case(&below, under_below, under_below_expect).run();
+
+    let under_above = above.position().xy() + Vector2(30.0, 15.0);
+    let under_above_expect = Vector2(
+        below.position().x - gap_x - min_spacing,
+        above.position().y - gap_y - node_view::HEIGHT,
+    );
+    create_case(&above, under_above, under_above_expect).run();
+}
+
+struct InitialNodes {
+    above: graph_editor::Node,
+    below: graph_editor::Node,
+}
+
+impl InitialNodes {
+    fn obtain_from_graph_editor(graph_editor: &GraphEditor) -> Self {
+        let nodes = graph_editor.model.nodes.all.values();
+        let mut sorted = nodes.into_iter().sorted_by_key(|node| OrderedFloat(node.position().y));
+        match (sorted.next(), sorted.next()) {
+            (Some(below), Some(above)) => Self { above, below },
+            _ => panic!("Expected two nodes in initial Graph Editor"),
+        }
+    }
 }